Add subs, substitute and rename methods.
Substitute an expression with another one, possibly using regular
expression, in one equation, a subset of equations, or the whole model.

subs and substitute are interfaces for the substitution
method (introduced in this patch). The only difference is that the
substitute method works with regular expressions (which is often
safer).

If the substitution is limited to a symbol in the whole model, or in
all the equations where this symbol appears, the new method
rename (also introduced in this patch) is called.

 + New method issymbol, to test if a char array is a symbol in the model.
 + Small cosmetic fixes
 + Change subsref indexing behaviour. A model can now be accessed
   using the dot notation instead of parentheses to extract
   one equation. For example, model.k will return a modBuilder object
   containing only the equation for k. Previously, this could be
   achieved using model('k'), but it was not possible to access
   equations with names longer than one character.
 + Integration tests for the new methods.